chore: orbit-store: rename file/layout.rs → file/path_safety.rs [ORB-10408]
Agent-Model: claude-sonnet-5
