+++ crates/orbit-store/src/file/mod.rs
@@ -9,8 +9,8 @@
 pub(crate) mod adr_store;
 pub(crate) mod executor_def_store;
 pub(crate) mod friction_store;
-pub(crate) mod layout;
 pub(crate) mod learning_store;
+pub(crate) mod path_safety;
 pub(crate) mod policy_def_store;
 pub(crate) mod scoreboard;
 pub(crate) mod skill_store;
